requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
discord.py==2.3.1
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.3.3
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
persiantools==3.0.1
protobuf==4.24.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyNaCl==1.5.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uritemplate==3.0.1
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
peewee==3.17.1